pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]

requires-python = ">=3.11.3"

name = "xy_django_app_information"

authors = [{ name = "helios", email = "yuyang.0515@qq.com" }]
description = "信息数据模型"

dynamic = ["version", "readme"]

classifiers = ["Programming Language :: Python :: 3.11.3"]

dependencies = [
    "django",
    "djangorestframework",
    "xy_type",
    "xy_string",
    "xy_django_model",
    "xy_admin_model_serializer",
]

[project.scripts]

[tool.setuptools.packages.find]
where = ["./"]
include = [
    "xy_django_app_information",
] # alternatively: `exclude = ["additional*"]`
namespaces = false

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = { file = "VERSION" }
readme = { file = ["README.rst", "USAGE.rst", "README.md", "USAGE.md"] }
